Add files via upload
this script is mostly useless to anyone but me (or hopefully people replicating my research).

it pulls data from the BCB SGS API and also merges with a "bank asset" variable in a .csv that i manually had to build for my academic research and is stored on my machine. 

the script then treats and transforms the data into a quarterly time series.